@@ -1,4 +1,9 @@
-GraduateStudents::Application.routes.draw do
+Rails.application.routes.draw do
+  # The priority is based upon order of creation: first created -> highest priority.
+  # See how all your routes lay out with "rake routes".
+
+  root :to => 'seminars#index'
+
   get "projects/destroy"
 
   get "observations/destroy"
@@ -12,7 +17,8 @@ GraduateStudents::Application.routes.draw do
   resources :participants, except: :new do
     get 'enroll', on: :member
     resource :transcript, only: :show
-    resources 'participant-activities', only: [:update, :destroy], as: :participant_activities, shallow: true
+    resources :activities, only: [:update, :destroy], as: :participant_activities, shallow:
+                                                                                                  true
     resources :observations, only: [:update, :destroy], shallow: true
     resources :projects, only: [:update, :destroy], shallow: true
   end
@@ -24,7 +30,7 @@ GraduateStudents::Application.routes.draw do
 
 
 
-  put 'admin/seminars/update_seminar' => 'admin/seminars#update_seminar', :as => 'update_whole_seminar'
+  patch 'admin/seminars/update_seminar' => 'admin/seminars#update_seminar', :as => 'update_whole_seminar'
 
   resources :seminars do
     resources :registrations, only: [:create, :destroy]
@@ -33,14 +39,14 @@ GraduateStudents::Application.routes.draw do
 
 
   resources :sessions, only: [:create, :destroy]
-  match 'login',  to: 'sessions#create'
-  match 'logout', to: 'sessions#destroy', via: [:get, :delete]
- 
+  get 'login',  to: 'sessions#create'
+  get 'logout', to: 'sessions#destroy', via: [:get, :delete]
+
 
   get 'search' => 'search#show', :as => 'search'
 
   get 'admin' => 'admin/seminars#index', :as => 'root_admin_url'
-   namespace :admin do
+  namespace :admin do
     # Directs /admin/seminars/* to Admin::SeminarController
     # (app/controllers/admin/seminars_controller.rb)
     resources :seminars do
@@ -50,22 +56,22 @@ GraduateStudents::Application.routes.draw do
     resources :spreadsheets, only: [:show]
     resources :settings, only: [:index, :update]
   end
-
-  #resources :registrations
   
 
-  # Sample of regular route:
-  #   match 'products/:id' => 'catalog#view'
-  # Keep in mind you can assign values other than :controller and :action
 
-  # Sample of named route:
-  #   match 'products/:id/purchase' => 'catalog#purchase', :as => :purchase
-  # This route can be invoked with purchase_url(:id => product.id)
+  # You can have the root of your site routed with "root"
+  # root 'welcome#index'
 
-  # Sample resource route (maps HTTP verbs to controller actions automatically):
+  # Example of regular route:
+  #   get 'products/:id' => 'catalog#view'
+
+  # Example of named route that can be invoked with purchase_url(id: product.id)
+  #   get 'products/:id/purchase' => 'catalog#purchase', as: :purchase
+
+  # Example resource route (maps HTTP verbs to controller actions automatically):
   #   resources :products
 
-  # Sample resource route with options:
+  # Example resource route with options:
   #   resources :products do
   #     member do
   #       get 'short'
@@ -77,37 +83,31 @@ GraduateStudents::Application.routes.draw do
   #     end
   #   end
 
-  # Sample resource route with sub-resources:
+  # Example resource route with sub-resources:
   #   resources :products do
   #     resources :comments, :sales
   #     resource :seller
   #   end
 
-  # Sample resource route with more complex sub-resources
+  # Example resource route with more complex sub-resources:
   #   resources :products do
   #     resources :comments
   #     resources :sales do
-  #       get 'recent', :on => :collection
+  #       get 'recent', on: :collection
   #     end
   #   end
 
-  # Sample resource route within a namespace:
+  # Example resource route with concerns:
+  #   concern :toggleable do
+  #     post 'toggle'
+  #   end
+  #   resources :posts, concerns: :toggleable
+  #   resources :photos, concerns: :toggleable
+
+  # Example resource route within a namespace:
   #   namespace :admin do
   #     # Directs /admin/products/* to Admin::ProductsController
   #     # (app/controllers/admin/products_controller.rb)
   #     resources :products
   #   end
-
-  # You can have the root of your site routed with "root"
-  # just remember to delete public/index.html.
-  #root :to => 'admin/seminars#index'
-  #root :to => 'registrations#new'
-  #root :to => 'welcome#index'
-  root :to => 'seminars#index'
-
-  # See how all your routes lay out with "rake routes"
-
-  # This is a legacy wild controller route that's not recommended for RESTful applications.
-  # Note: This route will make all actions in every controller accessible via GET requests.
-  # match ':controller(/:action(/:id))(.:format)'
 end
